requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dj-config-url==0.1.1
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
gunicorn==20.1.0
identify==2.5.17
isort==5.12.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.0.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
virtualenv==20.19.0
whitenoise==6.3.0
